Refactor mutations
